requirements.txt
streamlit
google-generativeai
pydantic
pandas
python-dotenv
typing-extensions
geopy